requirements.txt
django-sendgrid-webhook
pdf2image
pytesseract
pdfplumber
aiosmtplib==2.0.2
anyio==4.11.0
APScheduler==3.11.0
bcrypt==4.0.1
blinker==1.9.0
Brotli==1.1.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
cryptography==46.0.2
cssselect2==0.8.0
dnspython==2.8.0
ecdsa==0.19.1
email-validator==1.3.1
et_xmlfile==2.0.0
fastapi==0.118.0
fastapi-mail==1.3.1
fonttools==4.60.1
h11==0.16.0
httptools==0.6.4
idna==3.10
Jinja2==3.1.6
lxml==6.0.2
MarkupSafe==3.0.3
numpy==2.3.3
openpyxl==3.1.5
pandas==2.3.3
passlib==1.7.4
pillow==11.3.0
pyasn1==0.6.1
pycparser==2.23
pydantic==1.10.24
pydyf==0.11.0
pymongo==4.15.1
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
requests_pkcs12==1.27
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
starlette==0.48.0
tinycss2==1.4.0
tinyhtml5==2.0.0
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
uvicorn==0.37.0
uvloop==0.21.0
watchfiles==1.1.0
weasyprint==66.0
webencodings==0.5.1
websockets==15.0.1
xlsxwriter==3.2.9
zopfli==0.2.3.post1
